routertools: Use float for neighbor and gateway quality
This enables support for BATMAN V.

This requires changes to the MySQL database!

+++ ffmap/db/routers.py
@@ -135,7 +135,7 @@ mysql.execute("""
 	CREATE TABLE router_gw (
 		`router` mediumint(8) UNSIGNED NOT NULL,
 		`mac` char(17) COLLATE utf8_unicode_ci NOT NULL,
-		`quality` smallint(6) NOT NULL,
+		`quality` float NOT NULL,
 		`nexthop` char(17) COLLATE utf8_unicode_ci DEFAULT NULL,
 		`netif` varchar(15) COLLATE utf8_unicode_ci DEFAULT NULL,
 		`gw_class` varchar(25) COLLATE utf8_unicode_ci DEFAULT NULL,
@@ -166,7 +166,7 @@ mysql.execute("""
 		`router` mediumint(8) UNSIGNED NOT NULL,
 		`mac` char(17) COLLATE utf8_unicode_ci NOT NULL,
 		`netif` varchar(15) COLLATE utf8_unicode_ci NOT NULL,
-		`quality` smallint(6) NOT NULL,
+		`quality` float NOT NULL,
 		`type` varchar(10) COLLATE utf8_unicode_ci DEFAULT 'l2'
 	) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci
 """)
@@ -232,7 +232,7 @@ mysql.execute("""
 		`time` int(11) NOT NULL,
 		`router` mediumint(8) UNSIGNED NOT NULL,
 		`mac` char(17) COLLATE utf8_unicode_ci NOT NULL,
-		`quality` smallint(6) NOT NULL
+		`quality` float NOT NULL
 	) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci
 """)
 
@@ -247,7 +247,7 @@ mysql.execute("""
 		`time` int(11) NOT NULL,
 		`router` mediumint(8) UNSIGNED NOT NULL,
 		`mac` char(17) COLLATE utf8_unicode_ci NOT NULL,
-		`quality` smallint(6) NOT NULL
+		`quality` float NOT NULL
 	) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci
 """)
